installer: Remove "PHPBugTests" doc group
This was a pretty obscure menu item on doc.wikimedia.org at
<https://doc.wikimedia.org/mediawiki-core/master/php/>.

Merge this into the Installer group for now, which is where it's
located source-code wise. Whether this class is even needed still
is a separate matter. The PHP version is ancient, but there are
known one-to-many relations with C-libs, so I'll leave that to be
evaluated some other time.

(The Installer group is labeled "Deployment" which is its own
quirk that we should probably revisit as well.)

Change-Id: I00684840162989205c9e0feb4794b23aa8025b8f