Fix line selection in RandomImageGenerator
RandomImageGenerator::getRandomLines() is trying to select a random
subset of lines from a dictionary file, but the algorithm has a few bugs
that cause it to potentially leave entries in the returned array as null
and to be biased against the first few lines in the file.

This patch corrects the bugs to implement a Fisher–Yates (or Knuth)
shuffle,[1] run in reverse so that N does not need to be known ahead of
time and only maintaining the desired subset of the array in memory.

In local testing over many trials, it has produced output without nulls
100% of the time and the aggregate results do not have any obvious bias.

[1]: https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle

Bug: T222416
Change-Id: If136928bda2165198d3957a3536de85d465321fc