qunit: Remove div#mw-javascripttest-summary from SpecialJavaScriptTest
The intro text visually renders the same without it. The parsed
message already has its own block (a paragraph).

This div is a left-over from when the test runner was a skinned
page, and we had some CSS to style it in a fancy way. All that
was removed many years a go in favour of the 10-line standalone
HTML testrunner we have today.

Bug: T250045
Change-Id: Iffcca90857cf8e5df09a9287ab22fad109df7b79