resourceloader: Fix load.mock.php query parameter corruption in tests
=== Observe the bug

1. Run Special:JavaScriptTest
   (add ?module=mediawiki.loader to run only the relevant tests)
2. In the Network panel, check the JS requests to load.mock.php?…
3. Without this patch, they are like:
   "load.mock.php?1234?lang=en&modules=…&…"
   With this patch, they are like:
   "load.mock.php?lang=en&modules=…&…"

The question mark is only valid as the start of the query string,
not as divider between them. This means without this patch, the
"lang" parameter is simply ignored because it becomes part
of the key "1234?lang" with value "en".

=== What

The mock server doesn't do anything with "lang". And given that
RL sorts its query parameters for optimum cache-hit rate, the
corrupted parameter is always "lang", as its sorts before
"module" or "version", which our mock server does utilize.

As part of server-side compression of the startup module (d13e5b75),
we filter redundant base parameters that match the default. For
RLContext, this is `{ debug: false, lang: qqx, skin: fallback }`.

As such, if one were to mock the localisation backend with
uselang=qqx internally, the "lang" parameter will not need to be
sent, and thus the above bug will start corrupting the "modules"
paramater instead, which our test suite correctly detects as being
very badly broken.

=== Why

mediawiki.loader.test.js used QUnit.fixurl() as paranoid way to
avoid accidental caching. This blindly adds "?<random>" to the
url. Upstream QUnit assumes the URL will be a simple file on disk,
not expecting existing query parameters.

=== Fix

* Removing the call to QUnit.fixurl(). It was set by me years ago.
  But, there is no reason to believe a browser would cache this
  anyway. Plus, the file hardly ever changes. Just in case,
  set a no-cache header on the server side instead.

* Relatedly, the export of $VARS.reqBase is an associative array in
  PHP and becomes an object in JSON. Make sure this works even if
  the PHP array is empty, by casting to an object. Otherwise, it
  becomes `[]` instead of `{}` given an PHP php array is ambiguous
  in terms of whether it is meant as hashtable or list.

Bug: T250045
Change-Id: I3b8ff427577af9df3f1c26500ecf3646973ad34c