Add 'legacy' and 'i18n' features to ResourceLoaderSkinModule
The existing mediawiki.legacy styles module contains many legacy styles
as well as important i18n rules.

Begin migration of CSS rules from mediawiki.legacy to the
ResourceLoaderSkinModule, which can be tweaked by skins such as Minerva.

The mediawiki.legacy modules is updated to use this new 'legacy' feature
for backwards compatibility.

As agreed in the Frontend Standards Group, it is important we identify
the important rules and migrate them to this new module.

Bug: T169910
Bug: T217616
Change-Id: I1c1e2b912a41d29565e45e9e536c68ac46deb0e1