Drop MediaWikiServices::resetLanguageServices()
There is no reason this should be necessary. Resetting services should
do the trick, together with Language::$mLangObjCache while it exists.
Resetting only a subset of services is fragile and shouldn't be done,
because one of the reset services might later be injected into a service
that's not reset, resulting in bugs.

Change-Id: I8c72b466ba7337649f9256ee6c078c76163f9785