docs: Disable Doxygen parameter format warnings
Disable warnings for methods that are missing documentation,
or have invalid documentation.

While useful in theory, in practice this is almost entirely noise
that can be ignored. Doxygen falls back very well when it encounters
"invalid" documentation (e.g. render the type hint as plain text),
which is usually the right solution.

For example "Foo[]" means an array of "Foo" objects, but may
yield a warning for "unknown Foo[]" type. The fallback is to
not make a 404 link for "Foo[]", but render it as plain text
which seems desirable.

We already enforce validity in a better way with PHPCS and Phan,
which makes Doxygen's (limited and often incorrect) validation
obsolete.

Change-Id: I90000a4e33205e15ffabc8aeba0117615f1fd163