Save autoblock reasons with unexpanded templates
Autoblock reasons are built as a Message which takes the original
block's reason as a parameter. If the autoblock's reason is set using
the Message, it will be stored in the database with templates
expanded due to how a CommentStoreComment is constructed. Instead,
set the autoblock reason using a string with the unexpanded template.

Follow-up to 6a8920d3, which fixes how the reason is displayed in the
UI, but still expands the templates on saving.

Bug: T236501
Change-Id: I67a9941a9c039331576cbfc2ab58d9f20a7213e2