AccountCreation: Add user confirmation if reason is email address
The purpose of this is to prevent possible accidental outing caused by either
autofill or inattentive users when creating a new account while already logged
in. If the "reason" field is populated with an email address, it will prompt
the user to confirm that their intent is indeed to have an email address as
the create reason or allow them to change the reason. If they confirm or
change it from an email address to something else, then it will allow signup
to continue.

Bug: T233068
Change-Id: Ic80385924c84c5b6e029cef3cf589065907def30