resourceloader: Fix WikiModule cache invalidation order in Title::invalidateCache
Previously, the WANObjectCache key used by WikiModule was purged
by Title::invalidateCache from a db-precommit/idle callback.
This means it either runs right away (if there is no transaction
active) or from $dbw->commit().

This is a problem because the data that WikiModule is caching
relates directly to the 'page_touched' database field, which
Title::invalidateCache has not yet bumped at this point. Rather,
it does so later on, from a deferred update.

It's possible this has not yet caused an issue so far because
the WANObjectCache::touchCheckKey method leaves a tombstone
for 11 seconds (to counter race conditions and replication lag)
which might be enough for the current web request to finish
processing, and commit its transaction, and run any other deferred
update, and for Title's deferred update to bump page_touched,
and for that to be committed and replicated; all within 11s before
another web request interacts with WikiModule to repopulate the
cache key.

But, it's fragile at best and likely to get worse when we're
actually serving traffic multi-dc.

Change-Id: Icd9b3c11f60a829ea4c5d779eef3bebb643ebcf3