objectcache: make "db-replicated" use ObjectCache::newFromParams fully
The read/write BagOStuff sub-instances now use this method.

This makes it possible to let ObjectCache handle dependency injection
of LoadBalancer information rather than having the constructor access
MediaWikiServices.

It also means that 'logger', 'reportDupes', and 'asyncHandler' have
the standard ObjectCache defaults. Note that 'keyspace' is also set,
though it was already set on the parent BagOStuff instance which is
actually used for makeKey() calls.

Change-Id: I522a35eb63862217d87023744e7106187eabb2a1