EditPage: Improve handling of missing revision contents
Historically, if the content of a revision was missing (e.g. a bad entry
in the `text` table), EditPage would display an empty textarea.
Ia94521b7 accidentally broke that, causing an uncaught exception.

This patch restores the previous behavior, with the addition of a notice
at the top of the page that the content couldn't be loaded.

This also cleans up the missing section handling so it isn't so easily
confused with other failures (and also so it doesn't pass false to a
method declared as taking Content|null).

Bug: T237570
Change-Id: Ia70de11c2e4833b202fde3028a1a94dfc741f0a5