Fix bad comparison of Title::getCascadeProtectionSources result
The method \Title::getCascadeProtectionSources() is – I'm sorry to say
that – insane in so far that it returns two entirely different results
depending on it's first, boolean $getPages parameter. The return value is
always a 2-element array. The first element is either an array of Titles,
or a boolean. The second element is either an array of arrays of
restriction levels, or a meaningless empty array.

This really does not belong in a single method, but in two.

However. The code touched in this patch gets a boolean, but compares it
like it is a number. Luckily this was not an issue because PHP's >
compares booleans just fine, by converting them to numbers first. Still
super confusing.

Change-Id: I6af9be8f075595452a515e72360756da9a707a4e