preferences: Display error message at the top of the form as well
When there's an error in one of the fields "below the fold" or in
another tab, it's hard notice that it wasn't saved, especially since
the success message is very subtle.

In normal forms, there's a red error message at the top (in addition
to messages next to invalid fields), which was missing from the
preferences form.

I ran into this while experimenting with signature field validation.

Change-Id: If365e35d15a52939397f2093b1b8f5113a62e22b