jobs: Unify a few not matching job constructor signatures
According to \Job::factory() it's not possible for any of the
parameters to be optional. The first is always a Title object, the
second is always an array.

For classes implementing GenericParameterJob the $title parameter is
gone and replaced with a 'title' element in the array. Same here:
the array is not optional.

Change-Id: Idb4ae2f5fb500956b156710dcfe623802c274a3e