Rewrite confusing loop in TestFileEditor helper class
I had to write a little test to understand what this loop is supposed
to do: All it really does is counting the number of consecutive empty
lines in the $this->lines array, starting at the offset $this->pos.

The only edge-case is when the array does not contain anything but empty
lines. In this case, the last empty line is skipped. This is because
explode( …, "\n\n\n" ) returns 4 elements, but we want to count the
number of newline characters, which is 3.

Change-Id: I74ae1ed6ba0a8d65a92a1dd07f681fd7ae4221b4