LogPager: Add IGNORE INDEX to avoid another MariaDB optimizer bug
This is very similar to I53a7ed59 (T223151). In that bug it was choosing
the `times` index when it should have used `actor_time` due to a
non-empty performer. This time it's choosing `times` when it should be
using `type_time` due to a non-empty set of types being queried.

Bug: T237026
Change-Id: I02c77dde6b5b54d7bea801135051b006d39459b0