hooks: Do not document the Parser object to be passed by ref
It does not make sense for hook handler functions to *expect* the
Parser to be passed by reference. Hook handlers not only can't do
anything with it being a reference, they should *not* use it to
replace the Parser with another one.

This is an obsolete relict from PHP 3 (?), where objects got passed as a
cloned copy, which was very expensive.

Note we can not update the Hooks::run() calls as long as a single
hook handler still *expects* a reference.

Bug: T193950
Change-Id: I5f9a3f56faec0e90a2839c064844928c3b5c9751