DifferenceEngine: Don't try counting revs between deleted revs with different page IDs
Recent changes have caused the method for counting the number of
revisions in between the two sides of a diff to throw an error when the
revisions have different page IDs.

This blows up some cases of trying to get a difference between deleted
revisions, as the page IDs on the deleted revisions might not match
and/or might not match the current ID returned by the Title object.

Bug: T237709
Change-Id: I415f7ebb57fa4e879396b0db0e3a79edc2880be5