WikiPage: Fix indeterminate ordering in doRollback
When multiple revisions have the same timestamp (as is the case from
WikiPageDbTestBase::testDoRollback(), for example), the query being used
is allowed to return them in any other. It'll probably do what we expect
on MySQL due to some internal details of how it uses indexes, but the
lack of deterministic ordering is why the test was failing in PostgreSQL
and why it failed for SQLite in I4817b868.

To make it deterministic, order by rev_id when rev_timestamp is equal.
This is what MySQL happens to do, what the test expects, and is most
likely to give us the right behavior.

Bug: T195807
Change-Id: I8bb07fb6d0ff8f250b36ccc908429f264d52cd34