WikiPage: Fix indeterminate ordering in archiveRevisions
When multiple revisions have the same timestamp (as is the case from
ApiDeleteTest::testBatchedDelete(), for example), the query being used
is allowed to return them in any other. It'll probably do what we expect
on MySQL due to some internal details of how it uses indexes, but the
lack of deterministic ordering is why the test was failing in PostgreSQL
in I4817b868.

To make it deterministic, order by rev_id when rev_timestamp is equal.
This is what MySQL happens to do, what the batch-deletion job expects,
and is most likely to give us the right behavior.

Change-Id: If1d6529832260c5c0489f478568a85af5577ff44