PermissionManager: Remove long-obsolete comment
When cascading protection was originally added in r19095, the author
considered it too hard to correctly apply overlapping protection from
multiple source pages. A comment was placed in the code describing this
fact.

When r20471 a few months later did start applying the source-page
protections, it did not remove the comment that it made obsolete.

Change-Id: Ic8e115b501bd36f3bc7b89b77d4f8fa67e707838