rdbms: Have Database::makeWhereFrom2d assume $subKey is string-based
Until I70473280, integer literals were always quoted as strings, because
the databases we support all have no problem with casting
string-literals for comparisons and such.

But it turned out that gave MySQL/MariaDB's planner problems in some
queries, so we changed it to not quote actual PHP integers.

But then we run into the fact that PHP associative arrays don't preserve
the types of keys, it converts integer-like strings into actual
integers. And when those are passed to the DB unquoted for comparison
with a string-typed column, MySQL/MariaDB's planner has problems again
while PostgreSQL simply throws an error. Sigh.

This patch adjusts Database::makeWhereFrom2d to assume that the $subKey
column is going to need all values quoted, as is the case for all
callers in MediaWiki.

Bug: T239877
Change-Id: I69c125e8ab9e4d463eab35c6833aabdc436d7674