Avoid master connections for prop=info and intestactionsdetail=full
There is no security issue using RIGOR_FULL here, because the
information is not used to take an action. It is used for information to
the caller and the caller may not allow the action.
But even the caller allows the action, the action api code would check
permission with RIGOR_SECURE before doing the action.

Also use the constant to make clear which string is from api and which
one is from the PermissionManager

Bug: T239451
Change-Id: If182f0e967187704ba3fdd14592a0badff097571