resourceloader: Fix infinite loop in RLQ processing
RLQ.push( RLQ.shift() ); looks like it'd obviously be an infinite loop.
The reason it isn't is because we've overridden RLQ.push to execute its
argument. But it only does that if the argument is a function;
otherwise, it does add the argument to the array, causing an infinite
loop.

MediaWiki currently doesn't put non-function elements in RLQ early
enough for this code to run, but we want to be able to start doing that
without crashing everyone's browsers.

This was broken by df010dd28453e215f7a59fd731768c6f7eb60e0d. Go back to
putting the old RLQ value in a separate variable and resetting the
global RLQ to an empty array before processing the old array. That way
we're protected from an infinite loop, because the .push() and .shift()
operations happen on different arrays.

Change-Id: Idb10f5384f09e3cd681d417db28cca70d6e8dec2